move AbstractCustomView to it's own package
https://github.com/metasfresh/metasfresh/issues/5679
