+++ src/main/java/de/metas/ui/web/material/cockpit/stockdetails/StockDetailsView.java
@@ -2,9 +2,10 @@ package de.metas.ui.web.material.cockpit.stockdetails;
 
 import de.metas.handlingunits.model.I_M_HU_Stock_Detail_V;
 import de.metas.ui.web.document.filter.provider.DocumentFilterDescriptorsProvider;
-import de.metas.ui.web.view.AbstractCustomView;
 import de.metas.ui.web.view.IView;
 import de.metas.ui.web.view.ViewId;
+import de.metas.ui.web.view.template.AbstractCustomView;
+import de.metas.ui.web.view.template.IRowsData;
 import de.metas.ui.web.window.datatypes.DocumentId;
 import de.metas.util.Check;
