+++ src/main/java/de/metas/ui/web/order/products_proposal/view/ProductsProposalView.java
@@ -25,9 +25,9 @@ import de.metas.ui.web.order.products_proposal.model.ProductsProposalRow;
 import de.metas.ui.web.order.products_proposal.model.ProductsProposalRowAddRequest;
 import de.metas.ui.web.order.products_proposal.model.ProductsProposalRowChangeRequest;
 import de.metas.ui.web.order.products_proposal.model.ProductsProposalRowsData;
-import de.metas.ui.web.view.AbstractCustomView;
 import de.metas.ui.web.view.IEditableView;
 import de.metas.ui.web.view.ViewId;
+import de.metas.ui.web.view.template.AbstractCustomView;
 import de.metas.ui.web.window.datatypes.DocumentId;
 import de.metas.ui.web.window.datatypes.LookupValuesList;
 import de.metas.ui.web.window.datatypes.WindowId;
