+++ src/main/java/de/metas/ui/web/shipment_candidates_editor/ShipmentCandidateRows.java
@@ -16,9 +16,9 @@ import de.metas.inoutcandidate.api.ShipmentScheduleUserChangeRequest;
 import de.metas.inoutcandidate.api.ShipmentScheduleUserChangeRequestsList;
 import de.metas.ui.web.exceptions.EntityNotFoundException;
 import de.metas.ui.web.shipment_candidates_editor.ShipmentCandidateRowUserChangeRequest.ShipmentCandidateRowUserChangeRequestBuilder;
-import de.metas.ui.web.view.AbstractCustomView.IEditableRowsData;
-import de.metas.ui.web.view.AbstractCustomView.IRowsData;
 import de.metas.ui.web.view.IEditableView.RowEditingContext;
+import de.metas.ui.web.view.template.IEditableRowsData;
+import de.metas.ui.web.view.template.IRowsData;
 import de.metas.ui.web.window.datatypes.DocumentId;
 import de.metas.ui.web.window.datatypes.DocumentIdsSelection;
 import de.metas.ui.web.window.datatypes.json.JSONDocumentChangedEvent;
